requirements.txt
streamlit
pandas
google-cloud-bigquery
google-auth
db-dtypes